Receipt invoices + Inv Audit + delete cancel Fix across forms

+++ src/codebase/components/expenses/expenseEdit.js
@@ -90,7 +90,7 @@ function ExpenseEdit({ ID, operation, manualLoadData, setApiLoading, showSnackba
                 </IconButton>
                 <IconButton aria-label="Delete" title="Delete" color="error" onClick={() => {
                     const userInput = window.prompt("Type DELETE to confirm deletion of the item with ID: " + ID);
-                    if (userInput.toUpperCase() === "DELETE") {
+                    if (userInput && userInput.toUpperCase() === "DELETE") {
                         deleteItem();
                     } else {
                         // console.log("Delete operation cancelled");
