+++ src/codebase/components/expensetypes/expenseTypeEdit.js
@@ -77,7 +77,7 @@ function ExpenseTypeEdit({ ID, operation, manualLoadData, setApiLoading, showSna
                 </IconButton>
                 <IconButton aria-label="Delete" title="Delete" color="error" onClick={() => {
                     const userInput = window.prompt("Type DELETE to confirm deletion of the item with ID: " + ID);
-                    if (userInput.toUpperCase() === "DELETE") {
+                    if (userInput && userInput.toUpperCase() === "DELETE") {
                         deleteItem();
                     } else {
                         // console.log("Delete operation cancelled");
