+++ src/codebase/components/jobs/jobEdit.js
@@ -89,7 +89,7 @@ function JobEdit({ ID, operation, manualLoadData, setApiLoading, showSnackbar })
                 </IconButton>
                 <IconButton aria-label="Delete" title="Delete" color="error" onClick={() => {
                     const userInput = window.prompt("Type DELETE to confirm deletion of the item with ID: " + ID);
-                    if (userInput.toUpperCase() === "DELETE") {
+                    if (userInput && userInput.toUpperCase() === "DELETE") {
                         deleteItem();
                     } else {
                         // console.log("Delete operation cancelled");
