+++ src/codebase/components/jobtypes/jobTypeEdit.js
@@ -74,7 +74,7 @@ function JobTypeEdit({ ID, operation, manualLoadData, setApiLoading, showSnackba
                 </IconButton>
                 <IconButton aria-label="Delete" title="Delete" color="error" onClick={() => {
                     const userInput = window.prompt("Type DELETE to confirm deletion of the item with ID: " + ID);
-                    if (userInput.toUpperCase() === "DELETE") {
+                    if (userInput && userInput.toUpperCase() === "DELETE") {
                         deleteItem();
                     } else {
                         // console.log("Delete operation cancelled");
